# backend/booking_api/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel Booking API
  version: 0.1.0
  description: API for managing hotel bookings with user authentication and pricing calculations
servers:
  - url: http://localhost:9091
    description: Local development server
security:
  - JWTAuth: []
  - UserIdHeader: []
paths:
...
  /hotels/search:
    get:
      summary: Search Hotels
      description: Search mock hotels with optional filters
      operationId: search_hotels_hotels_search_get
      security: []
      parameters:
        - name: destination
          in: query
          schema:
            type: string
        - name: checkInDate
          in: query
          schema:
            type: string
            format: date
        - name: checkOutDate
          in: query
          schema:
            type: string
            format: date
        - name: guests
          in: query
          schema:
            type: integer
            minimum: 1
        - name: rooms
          in: query
          schema:
            type: integer
            minimum: 1
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_low, price_high, rating]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelSearchResponse'
        '500':
          description: Hotel search failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /hotels/resolve:
    get:
      summary: Resolve Hotel Id
      description: Resolve a hotel id from a name
      operationId: resolve_hotel_id_hotels_resolve_get
      security: []
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  hotelId:
                    type: string
                    nullable: true
                required:
                  - hotelId
        '500':
          description: Hotel resolve failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /hotels/{hotel_id}:
    get:
      summary: Get Hotel Details
      description: Retrieve hotel details and optional room data
      operationId: get_hotel_details_hotels__hotel_id__get
      security: []
      parameters:
        - name: hotel_id
          in: path
          required: true
          schema:
            type: string
        - name: checkInDate
          in: query
          schema:
            type: string
            format: date
        - name: checkOutDate
          in: query
          schema:
            type: string
            format: date
        - name: guests
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelDetailsResponse'
        '500':
          description: Hotel details failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /hotels/{hotel_id}/availability:
    get:
      summary: Check Hotel Availability
      description: Retrieve availability for a hotel
      operationId: get_hotel_availability_hotels__hotel_id__availability_get
      security: []
      parameters:
        - name: hotel_id
          in: path
          required: true
          schema:
            type: string
        - name: checkInDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: checkOutDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: guests
          in: query
          schema:
            type: integer
            minimum: 1
        - name: roomCount
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelAvailabilityResponse'
        '500':
          description: Hotel availability failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    get:
      summary: Get User Profile
      description: Retrieve the authenticated user's profile information
      operationId: get_profile_auth_profile_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Authentication Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings:
    get:
      summary: Get All Bookings
      description: Retrieve all bookings for the authenticated user
      operationId: get_bookings_bookings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          description: Authentication Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Storage Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create Booking
      description: Create a new hotel booking
      operationId: create_booking_bookings_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking Created Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBookingResponse'
        '401':
          description: Authentication Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Booking Persistence Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{booking_id}:
    get:
      summary: Get Booking Details
      description: Retrieve details of a specific booking
      operationId: get_booking_bookings__booking_id__get
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^BK[A-F0-9]{8}$'
            example: BK12345678
          description: The unique booking identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401':
          description: Authentication Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Storage Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update Booking
      description: Update an existing booking
      operationId: update_booking_bookings__booking_id__put
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^BK[A-F0-9]{8}$'
            example: BK12345678
          description: The unique booking identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookingRequest'
      responses:
        '200':
          description: Booking Updated Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateBookingResponse'
        '401':
          description: Authentication Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Booking Update Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Cancel Booking
      description: Cancel an existing booking
      operationId: cancel_booking_bookings__booking_id__delete
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^BK[A-F0-9]{8}$'
            example: BK12345678
          description: The unique booking identifier
      responses:
        '200':
          description: Booking Cancelled Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelBookingResponse'
        '401':
          description: Authentication Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Booking Cancel Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    JWTAuth:
      type: apiKey
      in: header
      name: x-jwt-assertion
      description: JWT token for authentication
    UserIdHeader:
      type: apiKey
      in: header
      name: x-user-id
      description: User ID header for authentication (alternative to JWT)
  schemas:
    HotelSummary:
      type: object
      description: Summary hotel data from mock dataset
      additionalProperties: true
...
    HotelSearchResponse:
      type: object
      properties:
        hotels:
          type: array
          items:
            $ref: '#/components/schemas/HotelSummary'
        metadata:
          $ref: '#/components/schemas/HotelSearchMetadata'
      required:
        - hotels
        - metadata
    HotelDetailsResponse:
      type: object
      properties:
        hotel:
          $ref: '#/components/schemas/HotelSummary'
        rooms:
          type: array
          items:
            type: object
            additionalProperties: true
        recentReviews:
          type: array
          items:
            type: object
            additionalProperties: true
        nearbyAttractions:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - hotel
        - rooms
        - recentReviews
        - nearbyAttractions
    HotelAvailabilityResponse:
      type: object
      properties:
        hotelId:
          type: string
        checkInDate:
          type: string
          format: date
        checkOutDate:
          type: string
          format: date
        availableRooms:
          type: array
          items:
            type: object
            additionalProperties: true
        totalAvailable:
          type: integer
      required:
        - hotelId
        - checkInDate
        - checkOutDate
        - availableRooms
        - totalAvailable
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: Authentication required
        errorCode:
          type: string
          description: Machine-readable error code
          example: AUTH_REQUIRED
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the error occurred
          example: '2024-01-27T10:30:00.000Z'
      required:
        - message
        - errorCode
        - timestamp
    UserProfile:
      type: object
      properties:
        userId:
          type: string
          description: Unique user identifier
          example: user123
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        firstName:
          type: string
          description: User's first name
          example: John
        lastName:
          type: string
          description: User's last name
          example: Doe
        phoneNumber:
          type: string
          description: User's phone number
          example: '+1234567890'
        profilePicture:
          type: string
          format: uri
          description: URL to user's profile picture
          example: https://example.com/profile.jpg
        registrationDate:
          type: string
          format: date-time
          description: ISO 8601 timestamp of user registration
          example: '2024-01-01T00:00:00.000Z'
        userType:
          type: string
          enum:
            - GUEST
            - PREMIUM
          description: Type of user account
          example: PREMIUM
        authClaims:
          type: object
          description: Raw authentication claims from JWT
          additionalProperties: true
      required:
        - userId
...
    CreateBookingRequest:
      type: object
      properties:
        hotelId:
          type: string
          description: Unique hotel identifier
          example: hotel_123
        hotelName:
          type: string
          description: Name of the hotel
          example: Grand Plaza Hotel
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
          minItems: 1
          description: List of rooms to book
        userId:
          type: string
          description: User ID (optional if JWT provided)
          example: user123
        checkInDate:
          type: string
          format: date
          description: Check-in date in ISO 8601 format
          example: '2024-02-01'
        checkOutDate:
          type: string
          format: date
          description: Check-out date in ISO 8601 format
          example: '2024-02-04'
        numberOfGuests:
          type: integer
          minimum: 1
          description: Total number of guests
          example: 4
        primaryGuest:
          $ref: '#/components/schemas/PrimaryGuest'
        specialRequests:
          type: string
          description: Any special requests or notes
          example: Late check-in requested
      required:
        - hotelId
        - hotelName
        - rooms
        - checkInDate
        - checkOutDate
        - numberOfGuests
        - primaryGuest
    UpdateBookingRequest:
      type: object
      properties:
        hotelId:
          type: string
          description: Unique hotel identifier
          example: hotel_123
        hotelName:
          type: string
          description: Name of the hotel
          example: Grand Plaza Hotel
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
          description: List of rooms to book
        userId:
          type: string
          description: User ID (optional if JWT provided)
          example: user123
        checkInDate:
          type: string
          format: date
          description: Check-in date in ISO 8601 format
          example: '2024-02-01'
        checkOutDate:
          type: string
          format: date
          description: Check-out date in ISO 8601 format
          example: '2024-02-04'
        numberOfGuests:
          type: integer
          minimum: 1
          description: Total number of guests
          example: 4
        primaryGuest:
          $ref: '#/components/schemas/PrimaryGuest'
        specialRequests:
          type: string
          description: Any special requests or notes
          example: Late check-in requested
    Booking:
      type: object
      properties:
        bookingId:
          type: string
          pattern: '^BK[A-F0-9]{8}$'
          description: Unique booking identifier
          example: BK12345678
        hotelId:
          type: string
          description: Unique hotel identifier
          example: hotel_123
        hotelName:
          type: string
          description: Name of the hotel
          example: Grand Plaza Hotel
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
          description: List of booked rooms
        userId:
          type: string
          description: User ID who made the booking
          example: user123
        checkInDate:
          type: string
          format: date
          description: Check-in date
          example: '2024-02-01'
        checkOutDate:
          type: string
          format: date
          description: Check-out date
          example: '2024-02-04'
        numberOfGuests:
          type: integer
          description: Total number of guests
          example: 4
        primaryGuest:
          $ref: '#/components/schemas/PrimaryGuest'
        pricing:
          type: array
          items:
            $ref: '#/components/schemas/Pricing'
          description: Pricing details for the booking
        bookingStatus:
          type: string
          enum:
            - CONFIRMED
            - CANCELLED
          description: Current status of the booking
          example: CONFIRMED
        bookingDate:
          type: string
          format: date-time
          description: ISO 8601 timestamp when booking was created
          example: '2024-01-27T10:30:00.000Z'
        confirmationNumber:
          type: string
          format: uuid
          description: Unique confirmation number
          example: CONFa1b2c3d4-e5f6-7890-abcd-ef1234567890
        specialRequests:
          type: string
          description: Any special requests or notes
          example: Late check-in requested
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp of last update
          example: '2024-01-27T11:00:00.000Z'
        cancelledAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when booking was cancelled
          example: '2024-01-27T12:00:00.000Z'
      required:
        - bookingId
        - userId
        - bookingStatus
        - bookingDate
        - confirmationNumber
    CreateBookingResponse:
      type: object
      properties:
        bookingId:
          type: string
          pattern: '^BK[A-F0-9]{8}$'
          description: Unique booking identifier
          example: BK12345678
        confirmationNumber:
          type: string
          format: uuid
          description: Unique confirmation number
          example: CONFa1b2c3d4-e5f6-7890-abcd-ef1234567890
        message:
          type: string
          description: Success message
          example: Booking confirmed successfully
        bookingDetails:
          $ref: '#/components/schemas/Booking'
      required:
        - bookingId
        - confirmationNumber
        - message
        - bookingDetails
    UpdateBookingResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: Booking updated successfully
        bookingDetails:
          $ref: '#/components/schemas/Booking'
      required:
        - message
        - bookingDetails
...